refactor large modules into smaller ones
