+++ crates/but-graph/src/init/types.rs
@@ -1,0 +1,436 @@
+use crate::{CommitFlags, CommitIndex, Edge, SegmentIndex};
+use petgraph::Direction;
+use petgraph::graph::EdgeReference;
+use petgraph::prelude::EdgeRef;
+use std::collections::{BTreeSet, VecDeque};
+use std::ops::Range;
+
+#[derive(Debug, Copy, Clone)]
+pub struct Limit {
+    inner: Option<usize>,
+    /// The commit we want to see to be able to assume normal limits. Until then there is no limit.
+    /// Each tracked commit is represented by bitflag, one for each goal, allowing commits to know
+    /// if they can be reached by the tracked commit.
+    /// The flag is empty if no goal is set.
+    goal: CommitFlags,
+}
+
+/// Lifecycle and builders
+impl Limit {
+    pub fn new(value: Option<usize>) -> Self {
+        Limit {
+            inner: value,
+            goal: CommitFlags::empty(),
+        }
+    }
+
+    /// Keep queueing without limit until `goal` is seen in a commit that has **it ahead of itself**.
+    /// Then stop searching for that goal.
+    /// `goals` are used to keep track of existing bitflags.
+    /// `origin` is used to know where the search for `goal` came from.
+    ///
+    /// ### Note
+    ///
+    /// No goal will be set if we can't track more goals, effectively causing traversal to stop earlier,
+    /// leaving potential isles in the graph.
+    /// This can happen if we have to track a lot of remotes, but since these are queued later, they are also
+    /// secondary and may just work for the typical remote.
+    pub fn with_indirect_goal(mut self, goal: gix::ObjectId, goals: &mut Goals) -> Self {
+        self.goal = goals.flag_for(goal).unwrap_or_default();
+        self
+    }
+
+    /// It's important to try to split the limit evenly so we don't create too
+    /// much extra gas here. We do, however, make sure that we see each segment of a parent
+    /// with one commit so we know exactly where it stops.
+    /// The problem with this is that we never get back the split limit when segments re-unite,
+    /// so effectively we loose gas here.
+    pub fn per_parent(&self, num_parents: usize) -> Self {
+        Limit {
+            inner: self
+                .inner
+                .map(|l| if l == 0 { 0 } else { (l / num_parents).max(1) }),
+            goal: self.goal,
+        }
+    }
+
+    /// Assure this limit won't perform any traversal after reaching its goals.
+    pub fn without_allowance(mut self) -> Self {
+        self.set_but_keep_goal(Limit::new(Some(0)));
+        self
+    }
+}
+
+/// Limit-check
+impl Limit {
+    /// Return `true` if this limit is depleted, or decrement it by one otherwise.
+    ///
+    /// `flags` are used to selectively decrement this limit.
+    /// Thanks to flag-propagation there can be no runaways.
+    pub fn is_exhausted_or_decrement(&mut self, flags: CommitFlags, next: &Queue) -> bool {
+        // Keep going if the goal wasn't seen yet, unlimited gas.
+        match self.goal_reachable(flags) {
+            Some(false) => return false,
+            Some(true) => self.set_goal_reached(),
+            None => {}
+        }
+        // Do not let *any* non-goal tip consume gas as long as there is still anything with a goal in the queue
+        // that need to meet their local branches.
+        // This is effectively only affecting the entrypoint tips, which isn't setup with a goal.
+        // TODO(perf): could we remember that we are a tip and look for our specific counterpart by matching the goal?
+        //             That way unrelated tips wouldn't cause us to keep traversing.
+        if self.goal_unset() && next.iter().any(|(_, _, _, limit)| !limit.goal_reached()) {
+            return false;
+        }
+        if self.inner.is_some_and(|l| l == 0) {
+            return true;
+        }
+        self.inner = self.inner.map(|l| l - 1);
+        false
+    }
+}
+
+/// Other access and mutation
+impl Limit {
+    /// Out-of-band way to use commit-flags differently - they never set the earlier flags, so we
+    /// can use them.
+    pub fn set_goal_reached(&mut self) {
+        self.goal.insert(CommitFlags::Integrated);
+    }
+
+    pub fn goal_reached(&self) -> bool {
+        self.goal_unset() || self.goal.contains(CommitFlags::Integrated)
+    }
+
+    fn goal_unset(&self) -> bool {
+        self.goal.is_empty()
+    }
+    /// Return `None` if this limit has no goal set, otherwise return `true` if `flags` contains it,
+    /// meaning it was reached through the commit the flags belong to.
+    /// This is useful to determine if a commit that is ahead was seen during traversal.
+    #[inline]
+    pub fn goal_reachable(&self, flags: CommitFlags) -> Option<bool> {
+        if self.goal_reached() {
+            None
+        } else {
+            Some(flags.contains(self.goal_flags()))
+        }
+    }
+
+    /// Return the goal flags, which may be empty.
+    pub fn goal_flags(&self) -> CommitFlags {
+        // Should only be one, at a time
+        let all_goals = self.goal.bits() & !CommitFlags::all().bits();
+        CommitFlags::from_bits_retain(all_goals)
+    }
+
+    /// Set our limit from `other`, but do not alter our goal.
+    pub fn set_but_keep_goal(&mut self, other: Limit) {
+        self.inner = other.inner;
+    }
+}
+
+/// Lifecycle
+impl Queue {
+    pub fn new_with_limit(limit: Option<usize>) -> Self {
+        Queue {
+            inner: Default::default(),
+            count: 0,
+            max: limit,
+        }
+    }
+}
+
+/// A queue to keep track of tips, which additionally counts how much was queued over time.
+pub struct Queue {
+    pub inner: VecDeque<QueueItem>,
+    /// The current number of queued items.
+    count: usize,
+    /// The maximum number of queuing operations, each representing one commit.
+    max: Option<usize>,
+}
+
+/// Counted queuing
+impl Queue {
+    #[must_use]
+    pub fn push_back_exhausted(&mut self, item: QueueItem) -> bool {
+        self.inner.push_back(item);
+        self.is_exhausted_after_increment()
+    }
+    #[must_use]
+    pub fn push_front_exhausted(&mut self, item: QueueItem) -> bool {
+        self.inner.push_front(item);
+        self.is_exhausted_after_increment()
+    }
+
+    fn is_exhausted_after_increment(&mut self) -> bool {
+        self.count += 1;
+        self.max.is_some_and(|l| self.count >= l)
+    }
+}
+
+/// Various other - good to know what we need though.
+impl Queue {
+    pub fn pop_front(&mut self) -> Option<QueueItem> {
+        self.inner.pop_front()
+    }
+    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut QueueItem> {
+        self.inner.iter_mut()
+    }
+    pub fn iter(&self) -> impl Iterator<Item = &QueueItem> {
+        self.inner.iter()
+    }
+}
+/// A set of commits to keep track of in bitflags.
+#[derive(Default)]
+pub struct Goals(Vec<gix::ObjectId>);
+
+impl Goals {
+    /// Return the bitflag for `goal`, or `None` if we can't track any more goals.
+    pub fn flag_for(&mut self, goal: gix::ObjectId) -> Option<CommitFlags> {
+        let existing_flags = CommitFlags::all().iter().count();
+        let max_goals = size_of::<CommitFlags>() * 8 - existing_flags;
+
+        let goals = &mut self.0;
+        let goal_index = match goals.iter().position(|existing| existing == &goal) {
+            None => {
+                let idx = goals.len();
+                goals.push(goal);
+                idx
+            }
+            Some(idx) => idx,
+        };
+        if goal_index >= max_goals {
+            tracing::warn!("Goals limit reached, cannot track {goal}");
+            None
+        } else {
+            Some(CommitFlags::from_bits_retain(
+                1 << (existing_flags + goal_index),
+            ))
+        }
+    }
+}
+
+#[derive(Debug, Copy, Clone)]
+pub enum Instruction {
+    /// Contains the segment into which to place this commit.
+    CollectCommit { into: SegmentIndex },
+    /// This is the first commit in a new segment which is below `parent_above` and which should be placed
+    /// at the last commit (at the time) via `at_commit`.
+    ConnectNewSegment {
+        parent_above: SegmentIndex,
+        at_commit: CommitIndex,
+    },
+}
+
+impl Instruction {
+    /// Returns any segment index we may be referring to.
+    pub fn segment_idx(&self) -> SegmentIndex {
+        match self {
+            Instruction::CollectCommit { into } => *into,
+            Instruction::ConnectNewSegment { parent_above, .. } => *parent_above,
+        }
+    }
+
+    pub fn with_replaced_sidx(self, sidx: SegmentIndex) -> Self {
+        match self {
+            Instruction::CollectCommit { into: _ } => Instruction::CollectCommit { into: sidx },
+            Instruction::ConnectNewSegment {
+                parent_above: _,
+                at_commit,
+            } => Instruction::ConnectNewSegment {
+                parent_above: sidx,
+                at_commit,
+            },
+        }
+    }
+}
+
+pub type QueueItem = (gix::ObjectId, CommitFlags, Instruction, Limit);
+
+#[derive(Debug)]
+pub(crate) struct EdgeOwned {
+    pub source: SegmentIndex,
+    pub target: SegmentIndex,
+    pub weight: Edge,
+    pub id: petgraph::graph::EdgeIndex,
+}
+
+impl From<EdgeReference<'_, Edge>> for EdgeOwned {
+    fn from(e: EdgeReference<'_, Edge>) -> Self {
+        EdgeOwned {
+            source: e.source(),
+            target: e.target(),
+            weight: *e.weight(),
+            id: e.id(),
+        }
+    }
+}
+
+/// A custom topological walk that always goes in a given [direction](Direction),
+/// does not need a graph reference, and which yields ranges of non-overlapping commit in a segment.
+///
+/// This walk assumes the worst-case graph where edges point to any commit, even from commits that
+/// aren't the first one or target commits that aren't the first one in a segment.
+///
+/// ### Note
+///
+/// In theory, a normal [`petgraph::visit::Topo`] would do here, if we assume that everything works
+/// as it should. So at some point, this code might be removed once it's clear we won't need it anymore.
+/// TODO: one fine day remove this in favor of `petgraph::visit::Topo`.
+pub struct TopoWalk {
+    /// The segment we
+    next: VecDeque<(SegmentIndex, Option<CommitIndex>)>,
+    /// Commits we have already yielded.
+    seen: gix::hashtable::HashSet,
+    /// If segments have no commits we can't identify them unless we track them separately.
+    seen_empty_segments: BTreeSet<SegmentIndex>,
+    /// In which direction to traverse.
+    direction: Direction,
+    /// If `true`, don't return the first segment which is always the starting point.
+    skip_tip: Option<()>,
+}
+
+/// Lifecycle
+impl TopoWalk {
+    /// Start a walk at `segment`, possibly only from `commit`.
+    pub fn start_from(
+        segment: SegmentIndex,
+        commit: Option<CommitIndex>,
+        direction: Direction,
+    ) -> Self {
+        TopoWalk {
+            next: {
+                let mut v = VecDeque::new();
+                v.push_back((segment, commit));
+                v
+            },
+            seen: Default::default(),
+            seen_empty_segments: Default::default(),
+            direction,
+            skip_tip: None,
+        }
+    }
+}
+
+/// Builder
+impl TopoWalk {
+    /// Call to not return the tip as part of the iteration.
+    pub fn skip_tip_segment(mut self) -> Self {
+        self.skip_tip = Some(());
+        self
+    }
+}
+
+/// Iteration
+impl TopoWalk {
+    /// Obtain the next segment and unseen commit range in the topo-walk.
+    /// Note that the returned range may yield an empty slice, or a sub-slice
+    /// of all available commits.
+    pub fn next(
+        &mut self,
+        graph: &crate::init::PetGraph,
+    ) -> Option<(SegmentIndex, Range<CommitIndex>)> {
+        while !self.next.is_empty() {
+            let res = self.next_inner(graph);
+            if res.is_some() {
+                if self.skip_tip.take().is_some() {
+                    continue;
+                }
+                return res;
+            }
+        }
+        None
+    }
+
+    fn next_inner(
+        &mut self,
+        graph: &crate::init::PetGraph,
+    ) -> Option<(SegmentIndex, Range<CommitIndex>)> {
+        let (segment, first_commit_index) = self.next.pop_front()?;
+        let available_range = self.select_range(graph, segment, first_commit_index)?;
+
+        for edge in graph.edges_directed(segment, self.direction) {
+            match self.direction {
+                Direction::Outgoing => {
+                    if edge
+                        .weight()
+                        .src
+                        .is_some_and(|src_cidx| !available_range.contains(&src_cidx))
+                    {
+                        continue;
+                    }
+                    self.next.push_back((edge.target(), edge.weight().dst));
+                }
+                Direction::Incoming => {
+                    if edge
+                        .weight()
+                        .dst
+                        .is_some_and(|dst_cidx| !available_range.contains(&dst_cidx))
+                    {
+                        continue;
+                    }
+                    self.next
+                        .push_back((edge.source(), edge.weight().src.map(|cidx| cidx + 1)));
+                }
+            }
+        }
+        Some((segment, available_range))
+    }
+
+    fn select_range(
+        &mut self,
+        graph: &crate::init::PetGraph,
+        segment: SegmentIndex,
+        first_commit_index: Option<CommitIndex>,
+    ) -> Option<Range<CommitIndex>> {
+        match first_commit_index {
+            None => {
+                debug_assert!(
+                    graph[segment].commits.is_empty(),
+                    "BUG: we always assure that we set the commit index if a segment has commits: {segment:?}"
+                );
+                if !self.seen_empty_segments.insert(segment) {
+                    return None;
+                }
+                Some(0..0)
+            }
+            Some(start_commit_idx) => {
+                let segment = &graph[segment];
+                let mut range = match self.direction {
+                    Direction::Outgoing => start_commit_idx..segment.commits.len(),
+                    Direction::Incoming => 0..start_commit_idx,
+                };
+
+                // NOTE: this assumes that we always consume all segments from a given starting position,
+                //       which is assured by the lines above that set the initial range.
+                let num_inserted_from_front = segment
+                    .commits
+                    .get(range.clone())?
+                    .iter()
+                    .take_while(|c| self.seen.insert(c.id))
+                    .count();
+                if num_inserted_from_front == 0 {
+                    let num_inserted_from_back = segment
+                        .commits
+                        .get(range.clone())?
+                        .iter()
+                        .rev()
+                        .take_while(|c| self.seen.insert(c.id))
+                        .count();
+                    if num_inserted_from_back == 0 {
+                        return None;
+                    }
+                    range.start = range.end - num_inserted_from_back;
+                } else {
+                    range.end = range.start + num_inserted_from_front;
+                }
+                debug_assert!(
+                    !range.is_empty(),
+                    "BUG: empty ranges should already have been handled, must return None"
+                );
+                Some(range)
+            }
+        }
+    }
+}
